Point to the NLX user documentation
If developers that want to build services on top of the NLX platform or network are reading this README they should learn as soon as possible that this is the documentation of the platform itself, not the documentation for the development of applications that use it.

Removed the `Documentation` section as it served this very purpose, whilst being confusing about the seperation between different kinds of documentation.