Explain the components in the repo
All components are maintained in a single repository so developers have all the tools and code to build and test the complete NLX network in a single repository. This simplifies version and dependency management and allows changes that affect multiple components to be combined in a single feature branch and pull-request.