Add the logo to the README
Since the logo is a wordmark of the letters NLX it has to also serve as the H1 heading.

Using some of the basic HTML allowed by most parsers including that of GitHub, I’ve been able to create a main heading that centers and is replaced by the logomark (when available).

For the width of the image I’ve divided it’s size by 4 so it scales well.

The image has been provided by @ehotting in the #design-system slack channel.