ci: disable stability days feature of Renovate bot
Reverts 8dd78f5d7857b7984172b70006e618fd8a7a9277

The stability days feature creates a new Pipeline in GitLab,
which suppresses the previous pipeline (containing the NLX jobs).
Therefore, it was possible to merge a MR without passing the tests.

This happened with
https://gitlab.com/commonground/nlx/nlx/-/merge_requests/1770/pipelines

There is an open issue for this on Renovates side, see
https://github.com/renovatebot/renovate/issues/6338