feat(management): move fetching application settings from PageTemplate to the App component
Fetching the settings in the PageTemplate made every page fetch the application settings.

This was also visible in our unit-tests, which were logging 'invalid JSON response' error
for every page. Because we did not mock the response for the GET /settings endpoint.

The `OrganizationInwachCheck` now only renders an alert if needed. While `App`
make sure the application settings are being loaded.